still not consolidating, but restructured slightly
